Testing ONC samples

+++ scripts/processing.py
@@ -2,6 +2,7 @@ from pathlib import Path
 import yaml
 import argparse
 from shutil import copy as sh_copy
+import pandas as pd
 from helpers import is_server_available, get_server_ip, load_config, setup_paths, check_mountpoint, check_rsync, \
     check_structure, check_docker_image, check_tso500_script, stage_object, process_object, transfer_results, get_queue
 from logging_ops import setup_logger, notify_bot
@@ -21,6 +22,7 @@ def main():
     with open('/mnt/NovaseqXplus/TSO_pipeline/01_Staging/pure-python-refactor/config.yaml', 'r') as file:
         config = yaml.safe_load(file)
         pipeline_dir: Path = Path(config['pipeline_dir'])
+        servers: list = config['available_servers']
 
     # TODO change in prod
     # if not is_server_available():
@@ -42,9 +44,20 @@ def main():
 
     path, input_type, _, tag, flowcell = queue.iloc[0]
 
-    is_last_sample = False
+    last_sample_queue = False
     if input_type == 'sample' and len(queue['Tag'][queue['Tag'] == tag]) == 1:
-        is_last_sample = True
+        last_sample_queue = True
+
+    last_sample_run = False
+    queues = []
+    for server in servers:
+        queue_file = pipeline_dir.parent.parent / f'{server}_QUEUE.txt'
+        pending_file = pipeline_dir.parent.parent / f'{server}_PENDING.txt'
+        queues.append(get_queue(pending_file=pending_file, queue_file=queue_file))
+    queue_merged = pd.concat(queues, ignore_index=True)
+    if len(queue_merged['Tag'][queue_merged['Tag'] == tag]) == 1:
+        last_sample_run = True
+
 
     config = load_config('/mnt/NovaseqXplus/TSO_pipeline/01_Staging/pure-python-refactor/config.yaml')
 
@@ -53,24 +66,22 @@ def main():
     logger = setup_logger(logger_name='Logger',log_file=paths['log_file'])
 
     check_mountpoint(paths=paths, logger=logger)
-
     check_structure(paths=paths, logger=logger)
-
     check_docker_image(logger=logger)
-
     check_rsync(paths=paths, logger=logger)
-
     check_tso500_script(paths=paths, logger=logger)
 
-    paths['analyzing_tag'].touch()
-    paths['queued_tag'].unlink()
-    stage_object(paths=paths, input_type=input_type, is_last_sample=is_last_sample, logger=logger)
 
-    process_object(paths=paths, input_type=input_type, is_last_sample=is_last_sample, logger=logger)
+    if not paths['analyzing_tag'].exists:
+        paths['analyzing_tag'].touch()
+        paths['queued_tag'].unlink()
+    stage_object(paths=paths, input_type=input_type, last_sample_queue=last_sample_queue, logger=logger)
 
-    transfer_results(paths=paths, input_type=input_type, is_last_sample=is_last_sample, logger=logger, testing=testing)
+    process_object(paths=paths, input_type=input_type, last_sample_queue=last_sample_queue, logger=logger)
 
-    if is_last_sample or input_type == 'run':
+    transfer_results(paths=paths, input_type=input_type, last_sample_queue=last_sample_queue, logger=logger, testing=testing)
+
+    if last_sample_run or input_type == 'run':
         paths['analyzed_tag'].touch()
         paths['analyzing_tag'].unlink()
